Modularized some code
