+++ lib/widgets/theme_utils.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+Color shade(BuildContext context) {
+  final theme = Theme.of(context);
+  return theme.brightness == Brightness.light
+      ? Color.fromARGB(181, 52, 78, 70)
+      : Color(0xFF344E46);
+}
+
+Color unshade(BuildContext context) {
+  final theme = Theme.of(context);
+  return theme.brightness == Brightness.light
+      ? Color.fromARGB(188, 29, 44, 31)
+      : Color(0xFF1D2C1F);
+}
